Former-commit-id: 7f12439aade073979a8318976a2bf88b59c45180

+++ BackEnd/src/main/java/com/c206/backend/domain/history/dto/response/HistoryListResponseDto.java
@@ -24,7 +24,7 @@ public class HistoryListResponseDto {
 
     private Long petId;
 
-    private double distance;
+    private int distance;
 
     private int trashCount;
